+++ inst/test-suite/crochet-replace.R
@@ -3,68 +3,72 @@
 # 'replace' function behaves similarly to a regular R matrix.
 #
 # To run these tests as part your test suite for your custom type, prepare an
-# environment called 'CROCHET_REPLACE_ENV' that contains the following names:
-# 'CUSTOM_OBJECT' (an instance of your custom type that implements replacement
-# using the 'replace' function containing dummy data with at least 3 rows and 3
-# columns), 'COMPARE_OBJECT' (a regular matrix that represents the same dummy
-# data as your object), 'OUT_OF_BOUNDS_INT' (an integer that represents an out
-# of bounds value for your object in integer indexing), 'OUT_OF_BOUNDS_CHAR' (a
-# string that represents an out of bounds value for your object in character
-# indexing), 'VALUE_POOL' (a vector containing possible values that your custom
-# type can take on), and 'RESET' (a function that resets 'CUSTOM_OBJECT' and
+# environment that contains the following names: 'CUSTOM_OBJECT' (an instance
+# of your custom type that implements replacement using the 'replace' function
+# containing dummy data with at least 3 rows and 3 columns), 'COMPARE_OBJECT'
+# (a regular matrix that represents the same dummy data as your object),
+# 'OUT_OF_BOUNDS_INT' (an integer that represents an out of bounds value for
+# your object in integer indexing), 'OUT_OF_BOUNDS_CHAR' (a string that
+# represents an out of bounds value for your object in character indexing),
+# 'VALUE_POOL' (a vector containing possible values that your custom type can
+# take on), and 'RESET' (a function that resets 'CUSTOM_OBJECT' and
 # 'COMPARE_OBJECT' to their original state). Once the environment is prepared,
 # the tests can be sourced using 'source(system.file("test-suite",
-# "crochet-replace.R", package = "crochet"), local = TRUE)'.
+# "crochet-replace.R", package = "crochet"), local = env)' where 'env' is the
+# environment.
 #
 # See tests/testthat/test-stringmatrix.R, the BEDMatrix package, or the
 # LinkedMatrix package for examples.
 
+# Environment needs to be passed via 'local' argument of 'source'
+if (!exists("CROCHET_REPLACE_ENV", inherits = FALSE)) {
+
 # dims are the same
-expect_equal(dim(CROCHET_REPLACE_ENV$CUSTOM_OBJECT), dim(CROCHET_REPLACE_ENV$COMPARE_OBJECT))
+expect_equal(dim(CUSTOM_OBJECT), dim(COMPARE_OBJECT))
 
 # dimnames are the same
-expect_equal(dimnames(CROCHET_REPLACE_ENV$CUSTOM_OBJECT), dimnames(CROCHET_REPLACE_ENV$COMPARE_OBJECT))
+expect_equal(dimnames(CUSTOM_OBJECT), dimnames(COMPARE_OBJECT))
 
 # minimum dimension requirement are met
-expect_true(nrow(CROCHET_REPLACE_ENV$CUSTOM_OBJECT) > 3)
-expect_true(ncol(CROCHET_REPLACE_ENV$CUSTOM_OBJECT) > 3)
-expect_true(length(CROCHET_REPLACE_ENV$CUSTOM_OBJECT) > 3)
+expect_true(nrow(CUSTOM_OBJECT) > 3)
+expect_true(ncol(CUSTOM_OBJECT) > 3)
+expect_true(length(CUSTOM_OBJECT) > 3)
 
 # maximum dimension requirement are met
-expect_true(nrow(CROCHET_REPLACE_ENV$CUSTOM_OBJECT) < CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT)
-expect_true(ncol(CROCHET_REPLACE_ENV$CUSTOM_OBJECT) < CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT)
-expect_true(length(CROCHET_REPLACE_ENV$CUSTOM_OBJECT) < CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT)
+expect_true(nrow(CUSTOM_OBJECT) < OUT_OF_BOUNDS_INT)
+expect_true(ncol(CUSTOM_OBJECT) < OUT_OF_BOUNDS_INT)
+expect_true(length(CUSTOM_OBJECT) < OUT_OF_BOUNDS_INT)
 
 # both matrices are the same
-expect_equal(CROCHET_REPLACE_ENV$CUSTOM_OBJECT[], CROCHET_REPLACE_ENV$COMPARE_OBJECT[])
+expect_equal(CUSTOM_OBJECT[], COMPARE_OBJECT[])
 
 test_replacement <- function(..., value) {
-    CROCHET_REPLACE_ENV$CUSTOM_OBJECT[...] <- value
-    CROCHET_REPLACE_ENV$COMPARE_OBJECT[...] <- value
-    expect_equal(CROCHET_REPLACE_ENV$CUSTOM_OBJECT[], CROCHET_REPLACE_ENV$COMPARE_OBJECT[], info = paste0("INFO: ", deparse(match.call())))
-    CROCHET_REPLACE_ENV$RESET()
+    CUSTOM_OBJECT[...] <- value
+    COMPARE_OBJECT[...] <- value
+    expect_equal(CUSTOM_OBJECT[], COMPARE_OBJECT[], info = paste0("INFO: ", deparse(match.call())))
+    RESET()
 }
 
 test_replacement_warning <- function(..., value) {
-    expect_warning(CROCHET_REPLACE_ENV$CUSTOM_OBJECT[...] <- value, info = paste0("INFO: ", deparse(match.call())))
-    expect_warning(CROCHET_REPLACE_ENV$COMPARE_OBJECT[...] <- value, info = paste0("INFO: ", deparse(match.call())))
-    expect_equal(CROCHET_REPLACE_ENV$CUSTOM_OBJECT[], CROCHET_REPLACE_ENV$COMPARE_OBJECT[], info = paste0("INFO: ", deparse(match.call())))
-    CROCHET_REPLACE_ENV$RESET()
+    expect_warning(CUSTOM_OBJECT[...] <- value, info = paste0("INFO: ", deparse(match.call())))
+    expect_warning(COMPARE_OBJECT[...] <- value, info = paste0("INFO: ", deparse(match.call())))
+    expect_equal(CUSTOM_OBJECT[], COMPARE_OBJECT[], info = paste0("INFO: ", deparse(match.call())))
+    RESET()
 }
 
 test_replacement_error <- function(..., value) {
-    expect_error(CROCHET_REPLACE_ENV$CUSTOM_OBJECT[...] <- value, info = paste0("INFO: ", deparse(match.call())))
-    expect_error(CROCHET_REPLACE_ENV$COMPARE_OBJECT[...] <- value, info = paste0("INFO: ", deparse(match.call())))
-    CROCHET_REPLACE_ENV$RESET()
+    expect_error(CUSTOM_OBJECT[...] <- value, info = paste0("INFO: ", deparse(match.call())))
+    expect_error(COMPARE_OBJECT[...] <- value, info = paste0("INFO: ", deparse(match.call())))
+    RESET()
 }
 
 test_replacement_not_implemented <- function(..., value) {
-    expect_error(CROCHET_REPLACE_ENV$CUSTOM_OBJECT[...] <- value, "not implemented", info = paste0("INFO: ", deparse(match.call())))
-    CROCHET_REPLACE_ENV$RESET()
+    expect_error(CUSTOM_OBJECT[...] <- value, "not implemented", info = paste0("INFO: ", deparse(match.call())))
+    RESET()
 }
 
-length <- length(CROCHET_REPLACE_ENV$CUSTOM_OBJECT)
-values <- CROCHET_REPLACE_ENV$VALUE_POOL
+length <- length(CUSTOM_OBJECT)
+values <- VALUE_POOL
 value <- values[1]
 
 # single replacement by nothing
@@ -113,19 +117,19 @@ test_replacement(TRUE, value = value)
 test_replacement(FALSE, value = value)
 test_replacement(c(TRUE, FALSE), value = value)
 test_replacement(c(FALSE, TRUE), value = value)
-m <- matrix(data = rnorm(length), nrow = nrow(CROCHET_REPLACE_ENV$CUSTOM_OBJECT), ncol = ncol(CROCHET_REPLACE_ENV$CUSTOM_OBJECT))
+m <- matrix(data = rnorm(length), nrow = nrow(CUSTOM_OBJECT), ncol = ncol(CUSTOM_OBJECT))
 test_replacement(m > 1, value = value)
 test_replacement(c(TRUE, NA), value = value)
 test_replacement(c(FALSE, NA), value = value)
 
 # single replacement by characters
-if (is.null(dimnames(CROCHET_REPLACE_ENV$CUSTOM_OBJECT))) {
+if (is.null(dimnames(CUSTOM_OBJECT))) {
     message("skipping character replacement because dimnames are NULL")
 } else {
-    ROW_NAME_1 <- rownames(CROCHET_REPLACE_ENV$CUSTOM_OBJECT)[1]
-    ROW_NAME_2 <- rownames(CROCHET_REPLACE_ENV$CUSTOM_OBJECT)[2]
-    COL_NAME_1 <- colnames(CROCHET_REPLACE_ENV$CUSTOM_OBJECT)[1]
-    COL_NAME_2 <- colnames(CROCHET_REPLACE_ENV$CUSTOM_OBJECT)[2]
+    ROW_NAME_1 <- rownames(CUSTOM_OBJECT)[1]
+    ROW_NAME_2 <- rownames(CUSTOM_OBJECT)[2]
+    COL_NAME_1 <- colnames(CUSTOM_OBJECT)[1]
+    COL_NAME_2 <- colnames(CUSTOM_OBJECT)[2]
     m <- matrix(data = c(ROW_NAME_1, COL_NAME_1, ROW_NAME_2, COL_NAME_2), ncol = 2, byrow = TRUE)
     test_replacement(m, value = value)
     m <- matrix(data = c(ROW_NAME_2, COL_NAME_2, ROW_NAME_1, COL_NAME_1), ncol = 2, byrow = TRUE)
@@ -144,26 +148,26 @@ test_replacement(c(0, -1), value = value)
 test_replacement_error(c(0, 1, -1), value = value)
 
 # single out-of-bounds replacements
-if (!is.null(CROCHET_REPLACE_ENV$SKIP_OUT_OF_BOUNDS_TESTS) && CROCHET_REPLACE_ENV$SKIP_OUT_OF_BOUNDS_TESTS) {
+if (exists("SKIP_OUT_OF_BOUNDS_TESTS", inherits = FALSE) && SKIP_OUT_OF_BOUNDS_TESTS) {
     message("skipping out-of-bounds tests because tests were explicitly disabled")
 } else {
     # positive integers
-    test_replacement_not_implemented(CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT, value = value) # expansion behavior
-    test_replacement_not_implemented(c(CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT, 2), value = value) # expansion behavior
-    test_replacement_not_implemented(c(2, CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT), value = value) # expansion behavior
-    m <- matrix(data = c(CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT, CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT), ncol = 2, byrow = TRUE)
+    test_replacement_not_implemented(OUT_OF_BOUNDS_INT, value = value) # expansion behavior
+    test_replacement_not_implemented(c(OUT_OF_BOUNDS_INT, 2), value = value) # expansion behavior
+    test_replacement_not_implemented(c(2, OUT_OF_BOUNDS_INT), value = value) # expansion behavior
+    m <- matrix(data = c(OUT_OF_BOUNDS_INT, OUT_OF_BOUNDS_INT), ncol = 2, byrow = TRUE)
     test_replacement_error(m, value = value)
     # negative integers
-    test_replacement(-CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT, value = value)
-    test_replacement(c(-CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT, -2), value = value)
-    test_replacement(c(-2, -CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT), value = value)
+    test_replacement(-OUT_OF_BOUNDS_INT, value = value)
+    test_replacement(c(-OUT_OF_BOUNDS_INT, -2), value = value)
+    test_replacement(c(-2, -OUT_OF_BOUNDS_INT), value = value)
     # logicals
-    #test_replacement_not_implemented(rep_len(TRUE, CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT), value = value) # expansion behavior (tricky to implement)
-    #test_replacement_not_implemented(rep_len(FALSE, CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT), value = value) # expansion behavior (tricky to implement)
-    #test_replacement_not_implemented(rep_len(c(TRUE, FALSE), CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT), value = value) # expansion behavior (tricky to implement)
-    #test_replacement_not_implemented(rep_len(c(FALSE, TRUE), CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT), value = value) # expansion behavior (tricky to implement)
+    #test_replacement_not_implemented(rep_len(TRUE, OUT_OF_BOUNDS_INT), value = value) # expansion behavior (tricky to implement)
+    #test_replacement_not_implemented(rep_len(FALSE, OUT_OF_BOUNDS_INT), value = value) # expansion behavior (tricky to implement)
+    #test_replacement_not_implemented(rep_len(c(TRUE, FALSE), OUT_OF_BOUNDS_INT), value = value) # expansion behavior (tricky to implement)
+    #test_replacement_not_implemented(rep_len(c(FALSE, TRUE), OUT_OF_BOUNDS_INT), value = value) # expansion behavior (tricky to implement)
     # characters
-    m <- matrix(data = c(CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_CHAR, CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_CHAR), ncol = 2, byrow = TRUE)
+    m <- matrix(data = c(OUT_OF_BOUNDS_CHAR, OUT_OF_BOUNDS_CHAR), ncol = 2, byrow = TRUE)
     test_replacement_error(m, value = value)
 }
 
@@ -255,13 +259,13 @@ test_replacement(, c(FALSE, NA), value = value)
 test_replacement(c(FALSE, NA), c(FALSE, NA), value = value)
 
 # multi replacement by characters
-if (is.null(dimnames(CROCHET_REPLACE_ENV$CUSTOM_OBJECT))) {
+if (is.null(dimnames(CUSTOM_OBJECT))) {
     message("skipping character replacement because dimnames are NULL")
 } else {
-    ROW_NAME_1 <- rownames(CROCHET_REPLACE_ENV$CUSTOM_OBJECT)[1]
-    ROW_NAME_2 <- rownames(CROCHET_REPLACE_ENV$CUSTOM_OBJECT)[2]
-    COL_NAME_1 <- colnames(CROCHET_REPLACE_ENV$CUSTOM_OBJECT)[1]
-    COL_NAME_2 <- colnames(CROCHET_REPLACE_ENV$CUSTOM_OBJECT)[2]
+    ROW_NAME_1 <- rownames(CUSTOM_OBJECT)[1]
+    ROW_NAME_2 <- rownames(CUSTOM_OBJECT)[2]
+    COL_NAME_1 <- colnames(CUSTOM_OBJECT)[1]
+    COL_NAME_2 <- colnames(CUSTOM_OBJECT)[2]
     test_replacement(ROW_NAME_1, , value = value)
     test_replacement(, COL_NAME_1, value = value)
     test_replacement(ROW_NAME_1, COL_NAME_1, value = value)
@@ -300,52 +304,52 @@ test_replacement_error(, c(0, 1, -1), value = value)
 test_replacement_error(c(0, 1, -1), c(0, 1, -1), value = value)
 
 # multi out-of-bounds replacements
-if (!is.null(CROCHET_REPLACE_ENV$SKIP_OUT_OF_BOUNDS_TESTS) && CROCHET_REPLACE_ENV$SKIP_OUT_OF_BOUNDS_TESTS) {
+if (exists("SKIP_OUT_OF_BOUNDS_TESTS", inherits = FALSE) && SKIP_OUT_OF_BOUNDS_TESTS) {
     message("skipping out-of-bounds tests because tests were explicitly disabled")
 } else {
     # positive integers
-    test_replacement_error(CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT, , value = value)
-    test_replacement_error(, CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT, value = value)
-    test_replacement_error(CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT, CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT, value = value)
-    test_replacement_error(c(CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT, 2), , value = value)
-    test_replacement_error(, c(CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT, 2), value = value)
-    test_replacement_error(c(CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT, 2), c(CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT, 2), value = value)
-    test_replacement_error(c(2, CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT), , value = value)
-    test_replacement_error(, c(2, CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT), value = value)
-    test_replacement_error(c(2, CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT), c(2, CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT), value = value)
+    test_replacement_error(OUT_OF_BOUNDS_INT, , value = value)
+    test_replacement_error(, OUT_OF_BOUNDS_INT, value = value)
+    test_replacement_error(OUT_OF_BOUNDS_INT, OUT_OF_BOUNDS_INT, value = value)
+    test_replacement_error(c(OUT_OF_BOUNDS_INT, 2), , value = value)
+    test_replacement_error(, c(OUT_OF_BOUNDS_INT, 2), value = value)
+    test_replacement_error(c(OUT_OF_BOUNDS_INT, 2), c(OUT_OF_BOUNDS_INT, 2), value = value)
+    test_replacement_error(c(2, OUT_OF_BOUNDS_INT), , value = value)
+    test_replacement_error(, c(2, OUT_OF_BOUNDS_INT), value = value)
+    test_replacement_error(c(2, OUT_OF_BOUNDS_INT), c(2, OUT_OF_BOUNDS_INT), value = value)
     # negative integers
-    test_replacement(c(-CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT), , value = value)
-    test_replacement(, c(-CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT), value = value)
-    test_replacement(c(-CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT), c(-CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT), value = value)
-    test_replacement(c(-CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT, -2), , value = value)
-    test_replacement(, c(-CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT, -2), value = value)
-    test_replacement(c(-CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT, -2), c(-CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT, -2), value = value)
-    test_replacement(c(-2, -CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT), , value = value)
-    test_replacement(, c(-2, -CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT), value = value)
-    test_replacement(c(-2, -CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT), c(-2, -CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT), value = value)
+    test_replacement(c(-OUT_OF_BOUNDS_INT), , value = value)
+    test_replacement(, c(-OUT_OF_BOUNDS_INT), value = value)
+    test_replacement(c(-OUT_OF_BOUNDS_INT), c(-OUT_OF_BOUNDS_INT), value = value)
+    test_replacement(c(-OUT_OF_BOUNDS_INT, -2), , value = value)
+    test_replacement(, c(-OUT_OF_BOUNDS_INT, -2), value = value)
+    test_replacement(c(-OUT_OF_BOUNDS_INT, -2), c(-OUT_OF_BOUNDS_INT, -2), value = value)
+    test_replacement(c(-2, -OUT_OF_BOUNDS_INT), , value = value)
+    test_replacement(, c(-2, -OUT_OF_BOUNDS_INT), value = value)
+    test_replacement(c(-2, -OUT_OF_BOUNDS_INT), c(-2, -OUT_OF_BOUNDS_INT), value = value)
     # logicals
-    test_replacement_error(c(rep_len(TRUE, CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT)), , value = value)
-    test_replacement_error(, c(rep_len(TRUE, CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT)), value = value)
-    test_replacement_error(c(rep_len(TRUE, CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT)), c(rep_len(TRUE, CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT)), value = value)
-    test_replacement_error(c(rep_len(FALSE, CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT)), , value = value)
-    test_replacement_error(, c(rep_len(FALSE, CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT)), value = value)
-    test_replacement_error(c(rep_len(FALSE, CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT)), c(rep_len(FALSE, CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT)), value = value)
-    test_replacement_error(rep_len(c(TRUE, FALSE), CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT), , value = value)
-    test_replacement_error(, rep_len(c(TRUE, FALSE), CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT), value = value)
-    test_replacement_error(rep_len(c(TRUE, FALSE), CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT), rep_len(c(TRUE, FALSE), CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT), value = value)
-    test_replacement_error(rep_len(c(FALSE, TRUE), CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT), , value = value)
-    test_replacement_error(, rep_len(c(FALSE, TRUE), CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT), value = value)
-    test_replacement_error(rep_len(c(FALSE, TRUE), CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT), rep_len(c(FALSE, TRUE), CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_INT), value = value)
+    test_replacement_error(c(rep_len(TRUE, OUT_OF_BOUNDS_INT)), , value = value)
+    test_replacement_error(, c(rep_len(TRUE, OUT_OF_BOUNDS_INT)), value = value)
+    test_replacement_error(c(rep_len(TRUE, OUT_OF_BOUNDS_INT)), c(rep_len(TRUE, OUT_OF_BOUNDS_INT)), value = value)
+    test_replacement_error(c(rep_len(FALSE, OUT_OF_BOUNDS_INT)), , value = value)
+    test_replacement_error(, c(rep_len(FALSE, OUT_OF_BOUNDS_INT)), value = value)
+    test_replacement_error(c(rep_len(FALSE, OUT_OF_BOUNDS_INT)), c(rep_len(FALSE, OUT_OF_BOUNDS_INT)), value = value)
+    test_replacement_error(rep_len(c(TRUE, FALSE), OUT_OF_BOUNDS_INT), , value = value)
+    test_replacement_error(, rep_len(c(TRUE, FALSE), OUT_OF_BOUNDS_INT), value = value)
+    test_replacement_error(rep_len(c(TRUE, FALSE), OUT_OF_BOUNDS_INT), rep_len(c(TRUE, FALSE), OUT_OF_BOUNDS_INT), value = value)
+    test_replacement_error(rep_len(c(FALSE, TRUE), OUT_OF_BOUNDS_INT), , value = value)
+    test_replacement_error(, rep_len(c(FALSE, TRUE), OUT_OF_BOUNDS_INT), value = value)
+    test_replacement_error(rep_len(c(FALSE, TRUE), OUT_OF_BOUNDS_INT), rep_len(c(FALSE, TRUE), OUT_OF_BOUNDS_INT), value = value)
     # character
-    test_replacement_error(CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_CHAR, , value = value)
-    test_replacement_error(, CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_CHAR, value = value)
-    test_replacement_error(CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_CHAR, CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_CHAR, value = value)
-    test_replacement_error(c(ROW_NAME_1, CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_CHAR), , value = value)
-    test_replacement_error(, c(COL_NAME_1, CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_CHAR), value = value)
-    test_replacement_error(c(ROW_NAME_1, CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_CHAR), c(COL_NAME_1, CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_CHAR), value = value)
-    test_replacement_error(c(CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_CHAR, ROW_NAME_1), , value = value)
-    test_replacement_error(, c(CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_CHAR, COL_NAME_1), value = value)
-    test_replacement_error(c(CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_CHAR, ROW_NAME_1), c(CROCHET_REPLACE_ENV$OUT_OF_BOUNDS_CHAR, COL_NAME_1), value = value)
+    test_replacement_error(OUT_OF_BOUNDS_CHAR, , value = value)
+    test_replacement_error(, OUT_OF_BOUNDS_CHAR, value = value)
+    test_replacement_error(OUT_OF_BOUNDS_CHAR, OUT_OF_BOUNDS_CHAR, value = value)
+    test_replacement_error(c(ROW_NAME_1, OUT_OF_BOUNDS_CHAR), , value = value)
+    test_replacement_error(, c(COL_NAME_1, OUT_OF_BOUNDS_CHAR), value = value)
+    test_replacement_error(c(ROW_NAME_1, OUT_OF_BOUNDS_CHAR), c(COL_NAME_1, OUT_OF_BOUNDS_CHAR), value = value)
+    test_replacement_error(c(OUT_OF_BOUNDS_CHAR, ROW_NAME_1), , value = value)
+    test_replacement_error(, c(OUT_OF_BOUNDS_CHAR, COL_NAME_1), value = value)
+    test_replacement_error(c(OUT_OF_BOUNDS_CHAR, ROW_NAME_1), c(OUT_OF_BOUNDS_CHAR, COL_NAME_1), value = value)
 }
 
 # n-dimensional replacement causes an error
@@ -361,3 +365,5 @@ test_replacement(c(1, 2, 3), value = values[1:3])
 test_replacement_warning(c(1, 2, 3), value = values[1:2])
 test_replacement_warning(c(1, 2, 3), value = rep_len(values, length.out = length))
 test_replacement_warning(c(1, 2, 3), value = rep_len(values, length.out = length + 3))
+
+}